backend/pyproject.toml
[project]
name = "termsscope-analytics-backend"
version = "0.1.0"
description = "AI-powered Terms of Service analyzer backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    # LangGraph & LangChain
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    # LLM SDKs
    "openai>=1.50.0",
    "tiktoken>=0.7.0",
    # Chunking
    "chonkie>=0.3.0",
    # Web scraping
    "httpx>=0.27.0",
    "lxml>=5.0.0",
    # PDF parsing
    "PyMuPDF>=1.24.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg[binary,pool]>=3.2.0",
    # Redis
    "redis[hiredis]>=5.0.0",
    # Auth
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    # Config & Utils
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "scrapling[fetchers]>=0.4.7",
    "tenacity>=9.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]
evals = [
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "eval: LLM-as-judge eval tests (deselect with -m 'not eval')",
]
